+++ CHANGELOG.md
@@ -5,6 +5,18 @@
 格式基于[维护更新日志（Keep a Changelog）](https://keepachangelog.com/en/1.0.0/)，
 并且该项目遵循 [语义版本控制（Semantic Versioning）](https://semver.org/spec/v2.0.0.html)。
 
+
+## [1.0.2] - 2022-06-10
+
+- 给几个用到终端的命令的消息提示加了个跳转到输出面板的按钮
+
+## [1.0.1] - 2022-06-10
+
+- 修复 `TRGL: Play Video`以及 `TRGL: Export Video`命令的bug（这个bug很怪，目前还没想出来是怎么回事，改了之后怎么就可以运行了，It just work！）
+- 修复了会出现多个输出框的bug
+- 修复了会出现多个终端的bug
+- 在 `TRGL: Synthesized Speech`命令中添加了一个按钮消息弹框，语音合成完毕后可以手动打开合成好的`AsteriskMarkedLogFile.txt`（找不到监听终端运行结束指令的替代方案）
+
 ## [1.0.0] - 2022-06-09
 
 - 新增命令 `TRGL: Define Character`，可以根据已有的log文件自动生成角色配置表